Move command tests into a commands dir in test/
