+++ test/cli.mjs
