+++ test/commands/database/create.mjs
@@ -4,10 +4,10 @@ import { expect } from "chai";
 import { fql, ServiceError } from "fauna";
 import sinon from "sinon";
 
-import { run } from "../../src/cli.mjs";
-import { setupTestContainer as setupContainer } from "../../src/config/setup-test-container.mjs";
-import { AUTHENTICATION_ERROR_MESSAGE } from "../../src/lib/errors.mjs";
-import { mockAccessKeysFile } from "../helpers.mjs";
+import { run } from "../../../src/cli.mjs";
+import { setupTestContainer as setupContainer } from "../../../src/config/setup-test-container.mjs";
+import { AUTHENTICATION_ERROR_MESSAGE } from "../../../src/lib/errors.mjs";
+import { mockAccessKeysFile } from "../../helpers.mjs";
 
 describe("database create", () => {
   let container, logger, runQuery, accountAPI;
