+++ test/commands/query.mjs
@@ -4,17 +4,17 @@ import { expect } from "chai";
 import { NetworkError, ServiceError } from "fauna";
 import sinon from "sinon";
 
-import { run } from "../src/cli.mjs";
-import { setupTestContainer as setupContainer } from "../src/config/setup-test-container.mjs";
-import { NETWORK_ERROR_MESSAGE } from "../src/lib/errors.mjs";
-import { colorize } from "../src/lib/formatting/colorize.mjs";
-import { QUERY_INFO_CHOICES } from "../src/lib/options.mjs";
+import { run } from "../../src/cli.mjs";
+import { setupTestContainer as setupContainer } from "../../src/config/setup-test-container.mjs";
+import { NETWORK_ERROR_MESSAGE } from "../../src/lib/errors.mjs";
+import { colorize } from "../../src/lib/formatting/colorize.mjs";
+import { QUERY_INFO_CHOICES } from "../../src/lib/options.mjs";
 import {
   createV4QueryFailure,
   createV4QuerySuccess,
   createV10QueryFailure,
   createV10QuerySuccess,
-} from "./helpers.mjs";
+} from "../helpers.mjs";
 
 describe("query", function () {
   let container, logger, runQueryFromString;
